+++ libs/domains/variables/feature/src/index.ts
@@ -13,4 +13,4 @@ export * from './lib/variables-context/variables-context'
 export * from './lib/show-all-variables-toggle/show-all-variables-toggle'
 export * from './lib/code-editor-variables/code-editor-variables'
 export * from './lib/flow-create-variable/flow-create-variable'
-export * from './lib/input-suggestions-variables/input-suggestions-variables'
+export * from './lib/wrapper-suggestions-variables/wrapper-suggestions-variables'
